Rewrite SequenceServer::Blast.
A BLAST search is now evaluated in the context of runtime settings: location of
BLAST+ binaries and databases, and number of threads to use; provided by a
`SequenceServer` object initialized during startup.  Error encountered, if any,
when attempting a BLAST search are wrapped in an exception object which is
caught by `SequenceServer::WebBlast`, and propagated to the user.

Add a few tests.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>