Rewrite SequenceServer::Blast; change error handling mechanism slightly.
A BLAST search is now evaluated in the context of runtime settings: location of
BLAST+ binaries and databases, and number of threads to use; provided by a
`SequenceServer::Blast` object initialized during startup.

`SequenceServer::Blast#run` executes a BLAST search in the context of the
runtime settings provided by `self`, and returns a
`SequenceServer::Blast::Query` object if the search was successful.

Error encountered, if any, when attempting a BLAST search are wrapped in an
exception object which is caught by `SequenceServer::App`, and propagated to
the user.

`SequenceServer::Blast::Query` object knows what search parameters are
acceptable and what not, so we no longer require the `before` block in
`SequenceServer::App` to ensure incoming data is valid.  Checking search
parameters in the controller felt like it knew too much about what and how of
running a BLAST search, when its only focus should be to act as a HTTP layer on
top of the CLI wrapper.

In short, the `SequenceServer::Blast` is now a self contained 'module' with its
own responsibilities.

A slew of minor, related changes follow:

1. How server communicates error messages to the client: client now expects
HTML formatted error message from the server, instead of plain text.

2. `SequenceServer::App` is initialized per object.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

blast: introduce Databases class; provide algorithms, algorithm, commands, and command method

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

views: update for changes in Blast API

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

blast.get

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>