Request user to report 500 errors to admin or forum; remove not_found handler.
Any runtime error unless specifically handled is trapped as internal server
error (500) by SS and shown to the user with the full bactrace, requesting them
to report it to the admin of the SS installation or the public forum.

Removed not_found handler since it depended on 500.erb which I tweaked
specifically for 500 (internal server) errors.  Sinatra, in production, shows a
page with heading (and only a heading) "No Found" for 404 errors which is
almost as good as what we previously had :-[.  And we can install a more
helpful 400 page later.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>