depend on Sinatra 1.3.2
I don't really know if SequenceServer breaks b/w Sinatra 1.2.0 and 1.3.2.
But since I refer to the source code of the latest Sinatra release during
development, it is probably best to depend on the same.

Also, we are going to use features that would break in the next release of
Sinatra, so let's depend on 1.3.2 exact (instead of >= 1.3.2).

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>