client: show network, and server timeout errors
And request users to try again later (since it could be a temporary
network/maintenance issue too; see NOTE), or report to the admin or
forum, if problem persists.

Timeout or network errors previously appeared to the user as "a flicker of
spinner followed by no results or an empty modal", leaving them to figure out
for themselves what exactly happened.  Not good.

NOTE: Detecting timeout errors is tricky.  We can be sure that the response
will be empty if server timed-out.  But empty response can be caused due to
myriad of reasons, broadly referred to as "network issues" later.  Currently SS
assumes empty response to mean timeout or a temporary network issue, and
accordingly displays the error message.  We could try handling network errors
'better' in future.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>