Add hit-download-link towards the right margin of BLAST result (#70, #78).
Download works by passing "&download=true" to `get_sequences`.  SequenceServer
uses download icon from Bootstrap to indicate 'download FASTA'.

TODO: The link to download all-hits needs to be aligned with the rest.

I temporarily removed checking if the number of sequences retrieved is the same
as the number of sequences requested, and will add it back later.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>