Show hits in a modal.
Now that SequenceServer has a separate link to actually download the hits,
let's use the hit-hyperlink -- the one with sequence id (lcl|Hsal_00289) and
towards the left margin of BLAST results -- to open the link in a modal instead
of a new page.

Users can search through the sequence, and download the sequence and dismiss
the modal in one click if they found what they were looking for.  The modal
approach lets user focus on one hit a time (or all hits, if they choose to).
Plus we are able to provide better visuals, and a consistent interface with
minimum effort.

This consolidates SequenceServer completely into a one page app, closing #76.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>