Rewrite SequenceServer completely.
* Logger
* public and views -> app

bin/sequenceserver takes charge of scanning BLAST+ binaries and databases.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

Make database formatter work again.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

Introduce notion of 'runtime' and Server module.

SequenceServer is now a class and not a module.  An instance of SequenceServer
captures options derived from configuration file and makes available objects
that depend on them.  One can think of it as a kernel or a runtime that boots
up first and initializes and exposes available modules.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

Rewrite SequenceServer::Blast.

A BLAST search is now evaluated in the context of runtime settings: location of
BLAST+ binaries and databases, and number of threads to use; provided by a
`SequenceServer` object initialized during startup.  Error encountered, if any,
when attempting a BLAST search are wrapped in an exception object which is
caught by `SequenceServer::WebBlast`, and propagated to the user.

Add a few tests.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

Request user to report 500 errors to admin or forum; remove not_found handler.

Any runtime error unless specifically handled is trapped as internal server
error (500) by SS and shown to the user with the full bactrace, requesting them
to report it to the admin of the SS installation or the public forum.

Removed not_found handler since it depended on 500.erb which I tweaked
specifically for 500 (internal server) errors.  Sinatra, in production, shows a
page with heading (and only a heading) "No Found" for 404 errors which is
almost as good as what we previously had :-[.  And we can install a more
helpful 400 page later.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

client: show network, and server timeout errors

And request users to try again later (since it could be a temporary
network/maintenance issue too; see NOTE), or report to the admin or
forum, if problem persists.

Timeout or network errors previously appeared to the user as "a flicker of
spinner followed by no results or an empty modal", leaving them to figure out
for themselves what exactly happened.  Not good.

NOTE: Detecting timeout errors is tricky.  We can be sure that the response
will be empty if server timed-out.  But empty response can be caused due to
myriad of reasons, broadly referred to as "network issues" later.  Currently SS
assumes empty response to mean timeout or a temporary network issue, and
accordingly displays the error message.  We could try handling network errors
'better' in future.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

Organize error templates.

Error sub-view now resides in views/errors.erb, rendered and sent to the user
on the first request (get '/').  Moved 400.erb and 500.erb to views/errors, so
they are referred in the controller as 'errors/400' and 'errors/500'.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

Rename Database#hash to Database#id.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

Redo format_search_results (close #17, #70).

SequenceServer::Blast takes the full responsibility of running a BLAST search,
and making available data needed to construct an HTML report. WebBlast simply
pulls out data from the search object and renders an erb template to generate
the HTML formatted result.

It was too much work to completely get rid of HTML from the search object, so
some attributes return HTML / String that's slotted into the report template.
Eventually search object should act only as a data point to generate the search
report via erb.

External or internal hyperlinks (refs) are generated per hit object now and can
be applied without modifying the source tree (close #17).  See
example.config.rb.  This file should be included after requiring sequenceserver
in config.ru or elsewhere.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

Associate a Hit object with the database it came from (close #99).

[ap: integrate the commit after massive rewrite; accordingly change commit msg]
[ap: use `find` instead of `select` as there is only one database per hit]

Signed-off-by: Volodymyr Zavidovych <vo.zavidovych@gmail.com>
Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

fix bundler dependencies so seqserv works with bundler 1.3 (1.2+)

[ap: do not bump version]

close #108

Signed-off-by: Ben J Woodcroft <donttrustben@gmail.com>
Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

Switch to latest Sinatra and Rack.

Additionally depend on ~> x.y instead of ~> x.y.z.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

Pull refs for hits to the right so they all align in one line.

Except "Download FASTA of all hits". With the current report templates it's
difficult to align it with the rest of the refs, so it stays at its place.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

Show hits in a modal together with a download button.

This consolidates SequenceServer completely into a one page app, closing #76.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

fix parse seqid (this is broken at the moment)

TODO: do not display [FASTA] links if couldn't determine which database the hit
came from.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>

remove bundler

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>